requirements.txt
Flask==3.0.3
gunicorn
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.2
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.10
langchain-openai==0.1.7
langchain-groq==0.1.6
langchain-mistralai==0.1.9
langchain-huggingface==0.0.3
sentence-transformers==3.4.1
transformers==4.49.0
langchainhub==0.1.21
langgraph==0.1.1
faiss-cpu==1.10.0
rank_bm25==0.2.2
colorama==0.4.6
python-dotenv==1.0.1
pyvis==0.3.2
graphviz==0.20.3
grandalf==0.8
pandas
Pillow
regex
requests
SQLAlchemy==2.0.39
stqdm
tqdm
pdfplumber
pdf2image
mistralai==1.5.1
opencv-python==4.11.0.86 